Consolidate build logic and fix nightly worklflow
This commit updates the build logic for the apps, relying on a
common place for package version definition. This lets us update it
easier in the future. This also changes the nightly workflow to match
what it is in dapr/dapr as we should be able to pick up go version
changes automatically with it.

Signed-off-by: halspang <halspang@microsoft.com>