Update AKS Bicep to use Redis instead of CosmosDB. (#190)
We are facing some issues using actor with a cosmosDB state store.
This issue might be related to dapr/dapr#6339. We are moving the setup
to use Redis instead of CosmosDB as our state store. This matches
our current longhaul setup.

While this might seem in contradiction with dapr/components-contrib#2071
and dapr/cli#1328, unblocking this issue will allow for
easier and predictable reproductions of our longhaul setup. We might
revisit the use of CosmosDB as a state store in the future.

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>