Release longhaul bounces app on update.
The Github action that updates the nightly/weekly longhaul updates dapr
version **and** does a rollout restart deploy on the apps. The action
that updates the release longhaul **doesn't**. This commit adds the same
logic from weekly to the release longhaul to also bounce the apps
running on the cluster and update their dapr version.

Signed-off-by: Tiago Alves Macambira <tmacam@burocrata.org>